refactor(codegen): split rust_codegen_util.rs into per-concern submodules
Split the 2,641-line codegen/rust_codegen_util.rs into a directory module,
mirroring the spec/chumsky_{adapter,parser} and adapt/pinocchio_profile
splits. Move-only: no behavior or generated-output changes (1006 tests + all
four snapshot suites byte-identical; the 13 emitted-Rust raw-string blocks
moved verbatim, no regex touched their contents).

- guards.rs   guard translation + top-level && split
- pubkey.rs   pubkey / Kani arithmetic rewriting + pubkey type helpers
- expr.rs     low-level string/comparison primitives (paren balance, cmp scan)
- property.rs property translation + arithmetic wrapping + solver selection
- effect.rs   account-env + value/state resolution + effect-triple lowering
- emit.rs     state/struct/enum/transition emission
- mod.rs      facade: module doc, shared imports, GuardTerm{,Source}, globs
- tests.rs    the #[cfg(test)] block, verbatim

The directory rename keeps crate::codegen::rust_codegen_util (and the root
re-export crate::rust_codegen_util) intact; mod.rs re-exports each submodule
`pub(crate) use <sub>::*` so all crate::rust_codegen_util::<name> call sites
across the 11 consumer modules resolve unchanged. Six private fns referenced
across submodule boundaries were bumped to pub(crate) (fix_equality_operator,
wrap_arithmetic, find_next_equality_op, find_cmp_lhs_start, find_cmp_rhs_end,
contains_whole_word, mentions_handler_account_pubkey); bucket-internal helpers
stayed private. No include_str! in this file, so no path adjustments.
