--- crates/qedgen/src/codegen/rust_codegen_util.rs
@@ -1,2641 +1,0 @@
-/// Shared helpers for generating Rust code from qedspec IR.
-///
-/// Used by both `proptest_gen` and `kani` to avoid duplicating
-/// the qedspec-to-Rust translation logic.
-use crate::check::{ParsedHandler, ParsedProperty, ParsedSpec};
-use crate::codegen_shared::DslTypeExt;
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum GuardTermSource {
-    Guard,
-    Requires { error_name: Option<String> },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct GuardTerm {
-    pub source: GuardTermSource,
-    pub rust_expr: String,
-}
-
-/// Translate a qedspec guard expression to Rust syntax.
-///
-/// Handles: state.field → s.field, Unicode operators → ASCII,
-/// Lean `=` equality → Rust `==`.
-pub fn translate_guard_to_rust(guard: &str, wrapping: bool) -> String {
-    let result = guard
-        .replace("state.", "s.")
-        .replace('≤', "<=")
-        .replace('≥', ">=")
-        .replace('∧', "&&")
-        .replace('∨', "||")
-        .replace('≠', "!=")
-        .replace(" and ", " && ")
-        .replace(" or ", " || ");
-    // Lean uses `=` for equality; Rust needs `==`. Replace standalone ` = `
-    // that isn't part of `<=`, `>=`, `!=`, or `==`.
-    let result = fix_equality_operator(&result);
-    if wrapping {
-        wrap_arithmetic(&result)
-    } else {
-        result
-    }
-}
-
-/// Rewrite handler-account pubkey references into a generated account
-/// environment. `foo.pubkey` becomes `<binder>.foo.pubkey`; `foo.key()` is
-/// normalized to the same pubkey field.
-pub fn rewrite_account_pubkey_refs(
-    expr: &str,
-    accounts: &[crate::check::ParsedHandlerAccount],
-    binder: &str,
-) -> String {
-    let mut out = expr.to_string();
-    for account in accounts {
-        let key_call = format!("{}.key()", account.name);
-        let pubkey_ref = format!("{}.pubkey", account.name);
-        let replacement = format!("{}.{}.pubkey", binder, account.name);
-        out = out.replace(&key_call, &replacement);
-        out = out.replace(&pubkey_ref, &replacement);
-    }
-    out
-}
-
-pub fn emit_kani_pubkey_helpers(out: &mut String) {
-    out.push_str("#[allow(dead_code)]\n");
-    out.push_str("fn pubkey_eq(a: &[u8; 32], b: &[u8; 32]) -> bool {\n");
-    out.push_str("    ");
-    for i in 0..32 {
-        if i > 0 {
-            out.push_str(" && ");
-        }
-        out.push_str(&format!("a[{i}] == b[{i}]"));
-    }
-    out.push_str("\n}\n\n");
-    out.push_str("#[allow(dead_code)]\n");
-    out.push_str("fn pubkey_ne(a: &[u8; 32], b: &[u8; 32]) -> bool {\n");
-    out.push_str("    !pubkey_eq(a, b)\n");
-    out.push_str("}\n\n");
-}
-
-pub fn spec_uses_pubkey(spec: &ParsedSpec) -> bool {
-    spec.state_fields
-        .iter()
-        .any(|(_, ty)| type_is_or_contains_pubkey(ty))
-        || spec.account_types.iter().any(|acct| {
-            acct.fields
-                .iter()
-                .any(|(_, ty)| type_is_or_contains_pubkey(ty))
-                || acct.variants.iter().any(|variant| {
-                    variant
-                        .fields
-                        .iter()
-                        .any(|(_, ty)| type_is_or_contains_pubkey(ty))
-                })
-        })
-        || spec.handlers.iter().any(|op| {
-            op.takes_params
-                .iter()
-                .chain(op.abstract_binders.iter())
-                .any(|(_, ty)| type_is_or_contains_pubkey(ty))
-        })
-}
-
-pub fn rewrite_kani_pubkey_comparisons(
-    expr: &str,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-) -> String {
-    let mut out = String::with_capacity(expr.len());
-    let mut cursor = 0;
-
-    while let Some((op_start, cmp)) = find_next_equality_op(expr, cursor) {
-        let op_end = op_start + cmp.len();
-        let lhs_start = find_cmp_lhs_start(expr, op_start);
-        let rhs_end = find_cmp_rhs_end(expr, op_end);
-
-        if lhs_start < cursor || rhs_end <= op_end {
-            out.push_str(&expr[cursor..op_end]);
-            cursor = op_end;
-            continue;
-        }
-
-        let lhs = expr[lhs_start..op_start].trim();
-        let rhs = expr[op_end..rhs_end].trim();
-        if kani_operand_is_pubkey(lhs, op, spec) && kani_operand_is_pubkey(rhs, op, spec) {
-            out.push_str(&expr[cursor..lhs_start]);
-            let helper = if cmp.trim() == "==" {
-                "pubkey_eq"
-            } else {
-                "pubkey_ne"
-            };
-            out.push_str(&format!("{helper}(&{lhs}, &{rhs})"));
-            cursor = rhs_end;
-        } else {
-            out.push_str(&expr[cursor..op_end]);
-            cursor = op_end;
-        }
-    }
-
-    out.push_str(&expr[cursor..]);
-    rewrite_kani_guard_arithmetic(&out)
-}
-
-pub fn rewrite_kani_guard_arithmetic(expr: &str) -> String {
-    let expr = rewrite_kani_bps_mul_div(expr);
-    rewrite_kani_checked_add_equality(&expr)
-}
-
-pub fn rewrite_kani_bps_mul_div(expr: &str) -> String {
-    let parenthesized = regex::Regex::new(
-        r"\((?P<a>[A-Za-z_][A-Za-z0-9_\.]*)\s*\*\s*(?P<b>[A-Za-z_][A-Za-z0-9_\.]*)\)\s*/\s*10000\b",
-    )
-    .expect("valid bps mul/div regex");
-    let rewritten = parenthesized
-        .replace_all(expr, "mul_bps_floor_u128($a, $b)")
-        .to_string();
-
-    let bare = regex::Regex::new(
-        r"\b(?P<a>[A-Za-z_][A-Za-z0-9_\.]*)\s*\*\s*(?P<b>[A-Za-z_][A-Za-z0-9_\.]*)\s*/\s*10000\b",
-    )
-    .expect("valid bare bps mul/div regex");
-    bare.replace_all(&rewritten, "mul_bps_floor_u128($a, $b)")
-        .to_string()
-}
-
-pub fn rewrite_kani_checked_add_equality(expr: &str) -> String {
-    let add_eq = regex::Regex::new(
-        r"\b(?P<a>[A-Za-z_][A-Za-z0-9_\.]*)\s*\+\s*(?P<b>[A-Za-z_][A-Za-z0-9_\.]*)\s*(?P<op>==|!=)\s*(?P<c>[A-Za-z_][A-Za-z0-9_\.]*|\d+)\b",
-    )
-    .expect("valid checked add equality regex");
-    add_eq
-        .replace_all(expr, "$a.checked_add($b) $op Some($c)")
-        .to_string()
-}
-
-pub fn spec_uses_kani_bps_mul_div_helper(spec: &ParsedSpec) -> bool {
-    let uses_helper = |expr: &str| rewrite_kani_bps_mul_div(expr) != expr;
-    spec.handlers.iter().any(|op| {
-        op.guard_str
-            .as_deref()
-            .map(|guard| uses_helper(&translate_guard_to_rust(guard, false)))
-            .unwrap_or(false)
-            || op.requires.iter().any(|req| uses_helper(&req.rust_expr))
-            || op
-                .aborts_if
-                .iter()
-                .any(|abort| uses_helper(&abort.rust_expr))
-            || op
-                .ensures
-                .iter()
-                .any(|ensures| uses_helper(&ensures.rust_expr_binary))
-            || op
-                .let_bindings
-                .iter()
-                .any(|(_, _, rust_expr)| uses_helper(rust_expr))
-    }) || spec
-        .properties
-        .iter()
-        .any(|property| property.rust_expression.as_deref().is_some_and(uses_helper))
-}
-
-pub fn negate_simple_top_level_comparison(expr: &str) -> Option<String> {
-    let trimmed = strip_balanced_outer_parens(expr.trim());
-    if contains_top_level_logical_op(trimmed) {
-        return None;
-    }
-    let bytes = trimmed.as_bytes();
-    let mut paren_depth = 0usize;
-    let mut bracket_depth = 0usize;
-    let mut brace_depth = 0usize;
-    let mut in_string = false;
-    let mut escaped = false;
-    let mut i = 0usize;
-
-    while i < bytes.len() {
-        let b = bytes[i];
-        if in_string {
-            if escaped {
-                escaped = false;
-            } else if b == b'\\' {
-                escaped = true;
-            } else if b == b'"' {
-                in_string = false;
-            }
-            i += 1;
-            continue;
-        }
-
-        match b {
-            b'"' => in_string = true,
-            b'(' => paren_depth += 1,
-            b')' => paren_depth = paren_depth.saturating_sub(1),
-            b'[' => bracket_depth += 1,
-            b']' => bracket_depth = bracket_depth.saturating_sub(1),
-            b'{' => brace_depth += 1,
-            b'}' => brace_depth = brace_depth.saturating_sub(1),
-            _ => {}
-        }
-
-        if paren_depth == 0 && bracket_depth == 0 && brace_depth == 0 {
-            for (op, negated) in [
-                ("==", "!="),
-                ("!=", "=="),
-                (">=", "<"),
-                ("<=", ">"),
-                (">", "<="),
-                ("<", ">="),
-            ] {
-                if trimmed[i..].starts_with(op) {
-                    let lhs = trimmed[..i].trim();
-                    let rhs = trimmed[i + op.len()..].trim();
-                    if !lhs.is_empty() && !rhs.is_empty() {
-                        return Some(format!("{lhs} {negated} {rhs}"));
-                    }
-                }
-            }
-        }
-        i += 1;
-    }
-    None
-}
-
-fn contains_top_level_logical_op(expr: &str) -> bool {
-    let bytes = expr.as_bytes();
-    let mut paren_depth = 0usize;
-    let mut bracket_depth = 0usize;
-    let mut brace_depth = 0usize;
-    let mut in_string = false;
-    let mut escaped = false;
-    let mut i = 0usize;
-
-    while i < bytes.len() {
-        let b = bytes[i];
-        if in_string {
-            if escaped {
-                escaped = false;
-            } else if b == b'\\' {
-                escaped = true;
-            } else if b == b'"' {
-                in_string = false;
-            }
-            i += 1;
-            continue;
-        }
-
-        match b {
-            b'"' => in_string = true,
-            b'(' => paren_depth += 1,
-            b')' => paren_depth = paren_depth.saturating_sub(1),
-            b'[' => bracket_depth += 1,
-            b']' => bracket_depth = bracket_depth.saturating_sub(1),
-            b'{' => brace_depth += 1,
-            b'}' => brace_depth = brace_depth.saturating_sub(1),
-            b'&' | b'|'
-                if i + 1 < bytes.len()
-                    && bytes[i + 1] == b
-                    && paren_depth == 0
-                    && bracket_depth == 0
-                    && brace_depth == 0 =>
-            {
-                return true;
-            }
-            _ => {}
-        }
-        i += 1;
-    }
-    false
-}
-
-fn strip_balanced_outer_parens(mut expr: &str) -> &str {
-    loop {
-        let trimmed = expr.trim();
-        if !(trimmed.starts_with('(') && trimmed.ends_with(')')) {
-            return trimmed;
-        }
-        let inner = &trimmed[1..trimmed.len() - 1];
-        if split_top_level_and(inner).len() == 1 && outer_parens_are_balanced(trimmed) {
-            expr = inner;
-        } else {
-            return trimmed;
-        }
-    }
-}
-
-fn outer_parens_are_balanced(expr: &str) -> bool {
-    let bytes = expr.as_bytes();
-    let mut depth = 0usize;
-    let mut in_string = false;
-    let mut escaped = false;
-    for (idx, b) in bytes.iter().copied().enumerate() {
-        if in_string {
-            if escaped {
-                escaped = false;
-            } else if b == b'\\' {
-                escaped = true;
-            } else if b == b'"' {
-                in_string = false;
-            }
-            continue;
-        }
-        match b {
-            b'"' => in_string = true,
-            b'(' => depth += 1,
-            b')' => {
-                depth = depth.saturating_sub(1);
-                if depth == 0 && idx + 1 < bytes.len() {
-                    return false;
-                }
-            }
-            _ => {}
-        }
-    }
-    depth == 0
-}
-
-fn find_next_equality_op(expr: &str, from: usize) -> Option<(usize, &'static str)> {
-    let eq = expr[from..].find(" == ").map(|p| (from + p, " == "));
-    let ne = expr[from..].find(" != ").map(|p| (from + p, " != "));
-    match (eq, ne) {
-        (Some(a), Some(b)) => Some(if a.0 <= b.0 { a } else { b }),
-        (Some(a), None) => Some(a),
-        (None, Some(b)) => Some(b),
-        (None, None) => None,
-    }
-}
-
-fn find_cmp_lhs_start(expr: &str, op_start: usize) -> usize {
-    let bytes = expr.as_bytes();
-    let mut i = op_start;
-    while i > 0 && bytes[i - 1].is_ascii_whitespace() {
-        i -= 1;
-    }
-
-    let mut depth = 0usize;
-    while i > 0 {
-        if depth == 0 && i >= 2 && (&expr[i - 2..i] == "&&" || &expr[i - 2..i] == "||") {
-            break;
-        }
-
-        let b = bytes[i - 1];
-        match b {
-            b')' => {
-                depth += 1;
-                i -= 1;
-            }
-            b'(' => {
-                if depth == 0 {
-                    break;
-                }
-                depth -= 1;
-                i -= 1;
-            }
-            b',' if depth == 0 => break,
-            _ => i -= 1,
-        }
-    }
-    i
-}
-
-fn find_cmp_rhs_end(expr: &str, op_end: usize) -> usize {
-    let bytes = expr.as_bytes();
-    let mut i = op_end;
-    while i < bytes.len() && bytes[i].is_ascii_whitespace() {
-        i += 1;
-    }
-
-    let mut depth = 0usize;
-    while i < bytes.len() {
-        if depth == 0 && i + 1 < bytes.len() && (&expr[i..i + 2] == "&&" || &expr[i..i + 2] == "||")
-        {
-            break;
-        }
-
-        match bytes[i] {
-            b'(' => {
-                depth += 1;
-                i += 1;
-            }
-            b')' => {
-                if depth == 0 {
-                    break;
-                }
-                depth -= 1;
-                i += 1;
-            }
-            b',' if depth == 0 => break,
-            _ => i += 1,
-        }
-    }
-    i
-}
-
-fn kani_operand_is_pubkey(operand: &str, op: &ParsedHandler, spec: &ParsedSpec) -> bool {
-    let operand = operand.trim().trim_start_matches('&').trim();
-    if operand.ends_with(".pubkey") || operand.ends_with(".key()") {
-        return true;
-    }
-
-    // `pubkey_eq` takes `&[u8; 32]`, so it only applies to a *scalar*
-    // pubkey operand: a `Pubkey` field, or an indexed element of a
-    // pubkey collection (`members[i]`). A bare reference to a whole
-    // collection (`s.members: [[u8; 32]; 32]`) must stay on `==`
-    // (derived array `PartialEq`) or it won't typecheck under Kani.
-    let indexed = operand.contains('[');
-
-    for prefix in ["s.", "pre.", "post."] {
-        if let Some(field) = operand.strip_prefix(prefix) {
-            return spec_field_is_scalar_pubkey(field, spec, indexed);
-        }
-    }
-
-    if let Some(field) = operand.strip_prefix("pre_") {
-        return spec_field_is_scalar_pubkey(field, spec, indexed);
-    }
-
-    if op
-        .takes_params
-        .iter()
-        .chain(op.abstract_binders.iter())
-        .any(|(name, ty)| name == operand && type_is_scalar_pubkey(ty))
-    {
-        return true;
-    }
-
-    spec_field_is_scalar_pubkey(operand, spec, indexed)
-}
-
-/// Resolve the declared spec type of a (possibly variant-prefixed,
-/// possibly subscripted) field path.
-fn spec_field_type<'a>(field: &str, spec: &'a ParsedSpec) -> Option<&'a str> {
-    let field = strip_variant_prefix_for_flat_state(field, spec);
-    let base = effect_target_base(field.as_str()).to_string();
-    if let Some((_, ty)) = spec.state_fields.iter().find(|(name, _)| *name == base) {
-        return Some(ty.as_str());
-    }
-    for acct in &spec.account_types {
-        if let Some((_, ty)) = acct.fields.iter().find(|(name, _)| *name == base) {
-            return Some(ty.as_str());
-        }
-        for variant in &acct.variants {
-            if let Some((_, ty)) = variant.fields.iter().find(|(name, _)| *name == base) {
-                return Some(ty.as_str());
-            }
-        }
-    }
-    None
-}
-
-/// True when the operand denotes a single `[u8; 32]` pubkey value —
-/// eligible for the `pubkey_eq`/`pubkey_ne` rewrite. A pubkey
-/// *collection* qualifies only when the operand indexes into it.
-fn spec_field_is_scalar_pubkey(field: &str, spec: &ParsedSpec, indexed: bool) -> bool {
-    match spec_field_type(field, spec) {
-        Some(ty) => type_is_scalar_pubkey(ty) || (indexed && type_is_or_contains_pubkey(ty)),
-        None => false,
-    }
-}
-
-fn type_is_or_contains_pubkey(ty: &str) -> bool {
-    ty.contains("Pubkey")
-}
-
-/// A single `Pubkey`, as opposed to a collection of them
-/// (`Map[N] Pubkey`, `[Pubkey; N]`, `Vec<Pubkey>`, …). Only scalar
-/// pubkeys lower to `[u8; 32]` and can be compared with `pubkey_eq`.
-fn type_is_scalar_pubkey(ty: &str) -> bool {
-    type_is_or_contains_pubkey(ty) && !type_is_pubkey_collection(ty)
-}
-
-fn type_is_pubkey_collection(ty: &str) -> bool {
-    let ty = ty.trim();
-    ty.contains('[') || ty.contains("Map") || ty.contains("Vec") || ty.contains("Array")
-}
-
-pub fn is_account_pubkey_ref(expr: &str, accounts: &[crate::check::ParsedHandlerAccount]) -> bool {
-    accounts
-        .iter()
-        .any(|a| expr == format!("{}.pubkey", a.name) || expr == format!("{}.key()", a.name))
-}
-
-pub fn handler_needs_account_env(op: &ParsedHandler) -> bool {
-    op.requires
-        .iter()
-        .any(|r| mentions_handler_account_pubkey(&r.rust_expr, &op.accounts))
-        || op
-            .guard_str
-            .as_ref()
-            .is_some_and(|g| mentions_handler_account_pubkey(g, &op.accounts))
-        || op
-            .effects
-            .iter()
-            .any(|(_, _, value)| is_account_pubkey_ref(value.trim(), &op.accounts))
-        || op.effect_branches.as_ref().is_some_and(|branches| {
-            branches.arms.iter().any(|arm| {
-                arm.effects
-                    .iter()
-                    .any(|(_, _, value)| is_account_pubkey_ref(value.trim(), &op.accounts))
-            })
-        })
-}
-
-pub fn handler_account_env_struct_name(op_name: &str) -> String {
-    let sanitized = crate::codegen_shared::sanitize_ident(op_name);
-    let mut out = String::new();
-    for part in sanitized.split('_').filter(|p| !p.is_empty()) {
-        let mut chars = part.chars();
-        if let Some(first) = chars.next() {
-            out.extend(first.to_uppercase());
-            out.push_str(chars.as_str());
-        }
-    }
-    if out.is_empty() {
-        "Handler".to_string()
-    } else {
-        format!("{}Accounts", out)
-    }
-}
-
-/// Translate a qedspec property expression to Rust.
-pub fn translate_property_to_rust(expr: &str, wrapping: bool) -> String {
-    let result = expr
-        .replace("state.", "s.")
-        .replace('≤', "<=")
-        .replace('≥', ">=")
-        .replace('∧', "&&")
-        .replace('∨', "||")
-        .replace('≠', "!=")
-        .replace(" and ", " && ")
-        .replace(" or ", " || ");
-    let result = fix_equality_operator(&result);
-    if wrapping {
-        wrap_arithmetic(&result)
-    } else {
-        result
-    }
-}
-
-/// Fix standalone ` = ` (Lean equality) to ` == ` (Rust equality),
-/// without touching compound operators like `<=`, `>=`, `!=`.
-fn fix_equality_operator(input: &str) -> String {
-    let mut safe = String::with_capacity(input.len());
-    let bytes = input.as_bytes();
-    let mut i = 0;
-    while i < bytes.len() {
-        if bytes[i] == b'='
-            && i > 0
-            && i + 1 < bytes.len()
-            && bytes[i - 1] == b' '
-            && bytes[i + 1] == b' '
-            && (i < 2 || (bytes[i - 2] != b'<' && bytes[i - 2] != b'>' && bytes[i - 2] != b'!'))
-            && (i + 2 >= bytes.len() || bytes[i + 1] != b'=')
-        {
-            safe.push_str("==");
-        } else {
-            safe.push(bytes[i] as char);
-        }
-        i += 1;
-    }
-    safe
-}
-
-/// Convert infix `a + b` and `a - b` to `a.wrapping_add(b)` and `a.wrapping_sub(b)`
-/// within comparison sub-expressions. Only transforms arithmetic within individual
-/// conjuncts/disjuncts — doesn't break boolean structure.
-fn wrap_arithmetic(expr: &str) -> String {
-    let parts: Vec<&str> = expr.split(" && ").collect();
-    let wrapped: Vec<String> = parts
-        .iter()
-        .map(|part| {
-            let sub_parts: Vec<&str> = part.split(" || ").collect();
-            sub_parts
-                .iter()
-                .map(|sub| wrap_arithmetic_atom(sub.trim()))
-                .collect::<Vec<_>>()
-                .join(" || ")
-        })
-        .collect();
-    wrapped.join(" && ")
-}
-
-fn wrap_arithmetic_atom(atom: &str) -> String {
-    for cmp in &[" <= ", " >= ", " < ", " > ", " == ", " != "] {
-        if let Some(pos) = atom.find(cmp) {
-            let lhs = &atom[..pos];
-            let rhs = &atom[pos + cmp.len()..];
-            let lhs_wrapped = wrap_arith_expr(lhs.trim());
-            let rhs_wrapped = wrap_arith_expr(rhs.trim());
-            return format!("{}{}{}", lhs_wrapped, cmp, rhs_wrapped);
-        }
-    }
-    atom.to_string()
-}
-
-fn wrap_arith_expr(expr: &str) -> String {
-    // Split on the RIGHTMOST top-level ` + ` / ` - ` and recurse on the LHS
-    // so chains lower left-associatively (`cap - used - reserved` →
-    // `cap.saturating_sub(used).saturating_sub(reserved)`), matching Lean's
-    // left-associative `Nat` subtraction. Subtraction is `saturating_sub`
-    // because Lean models guard/property arithmetic over `Nat` (`-`
-    // truncates at 0) — the predicate must test the same relation the Lean
-    // proof discharges. Addition keeps `wrapping_add`.
-    let bytes = expr.as_bytes();
-    let mut depth: i32 = 0;
-    let mut split: Option<(usize, u8)> = None;
-    let mut i = 0;
-    while i < bytes.len() {
-        match bytes[i] {
-            b'(' => depth += 1,
-            b')' => depth -= 1,
-            op @ (b'+' | b'-')
-                if depth == 0
-                    && i > 0
-                    && bytes[i - 1] == b' '
-                    && i + 1 < bytes.len()
-                    && bytes[i + 1] == b' ' =>
-            {
-                split = Some((i, op));
-            }
-            _ => {}
-        }
-        i += 1;
-    }
-    match split {
-        Some((pos, op)) => {
-            let lhs = wrap_arith_expr(expr[..pos].trim());
-            let rhs = expr[pos + 1..].trim();
-            if op == b'+' {
-                format!("{}.wrapping_add({})", lhs, rhs)
-            } else {
-                format!("{}.saturating_sub({})", lhs, rhs)
-            }
-        }
-        None => expr.to_string(),
-    }
-}
-
-/// For a field with an "add" effect, find its upper-bound field in property expressions.
-/// Property expressions are in Lean form (e.g. `s.approval_count ≤ s.member_count`).
-/// Returns the bounding field name if a `field ≤ bound` pattern is found.
-pub fn find_upper_bound_field(field: &str, properties: &[ParsedProperty]) -> Option<String> {
-    for prop in properties {
-        if let Some(ref expr) = prop.expression {
-            let norm = expr.replace('\u{2264}', "<=").replace('\u{2265}', ">=");
-            let field_pat = format!("s.{}", field);
-            if !norm.contains(&field_pat) && !norm.contains(field) {
-                continue;
-            }
-            for segment in norm.split("&&").chain(norm.split('\u{2227}')) {
-                let segment = segment.trim();
-                if let Some((lhs, rhs)) = segment.split_once("<=") {
-                    let lhs = lhs.trim();
-                    let rhs = rhs.trim();
-                    if lhs.ends_with(field) || lhs == format!("s.{}", field) {
-                        let bound = rhs
-                            .strip_prefix("s.")
-                            .or_else(|| rhs.strip_prefix("state."))
-                            .unwrap_or(rhs)
-                            .trim();
-                        if bound.chars().all(|c| c.is_alphanumeric() || c == '_')
-                            && !bound.is_empty()
-                            && !bound.chars().next().unwrap().is_ascii_digit()
-                        {
-                            return Some(bound.to_string());
-                        }
-                    }
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Emit assume statements for add effects with bounded properties.
-/// `assume_fmt` controls the output syntax, e.g.:
-///   - proptest: `"        prop_assume!(s.{field} < s.{bound}); // strict bound for add\n"`
-///   - kani:     `"    kani::assume(s.{field} < s.{bound}); // strict bound: {field} increments\n"`
-pub fn emit_add_strict_bounds(
-    out: &mut String,
-    op: &ParsedHandler,
-    properties: &[ParsedProperty],
-    assume_fmt: &str,
-) {
-    for (field, eff_op, _) in &op.effects {
-        if eff_op == "add" {
-            if let Some(bound) = find_upper_bound_field(field, properties) {
-                out.push_str(
-                    &assume_fmt
-                        .replace("{field}", field)
-                        .replace("{bound}", &bound),
-                );
-            }
-        }
-    }
-}
-
-/// Infer a Rust integer type from a constant's value magnitude.
-pub fn infer_const_type(value: &str) -> &'static str {
-    let clean_val = value.replace('_', "");
-    // Try unsigned first; fall through to signed only when a leading `-`
-    // rules out the unsigned path, picking the smallest type that fits.
-    if let Ok(v) = clean_val.parse::<u128>() {
-        if v <= u8::MAX as u128 {
-            "u8"
-        } else if v <= u16::MAX as u128 {
-            "u16"
-        } else if v <= u32::MAX as u128 {
-            "u32"
-        } else if v <= u64::MAX as u128 {
-            "u64"
-        } else {
-            "u128"
-        }
-    } else if let Ok(v) = clean_val.parse::<i128>() {
-        if v >= i8::MIN as i128 && v <= i8::MAX as i128 {
-            "i8"
-        } else if v >= i16::MIN as i128 && v <= i16::MAX as i128 {
-            "i16"
-        } else if v >= i32::MIN as i128 && v <= i32::MAX as i128 {
-            "i32"
-        } else if v >= i64::MIN as i128 && v <= i64::MAX as i128 {
-            "i64"
-        } else {
-            "i128"
-        }
-    } else {
-        "u64"
-    }
-}
-
-/// Pick a CBMC backend solver for a Kani effect-conformance harness based on
-/// the LHS field type and the RHS expression.
-///
-/// Returns the content of the `#[kani::solver(...)]` attribute (without the
-/// attribute wrapper). The three tiers:
-///
-/// * **cadical** — scalar / linear effects (no `*` or `/` reachable from the
-///   RHS). Default Kani solver; fast on bit-blasted boolean and linear-arith
-///   problems.
-/// * **minisat** — narrow-type multiplication/division (u8, u16, u32, bool).
-///   SAT-level solver that outperforms cadical on multiplication-heavy
-///   bit-blasts at narrow widths.
-/// * **bin = "z3"** — wide-type multiplication/division (u64, u128, i128).
-///   CBMC hands the problem to z3 as an SMT2 solver; z3's bit-vector theory
-///   handles nested `*`/`/` chains on 64+ bit types that SAT backends blow up
-///   on (the `amount * 125 / 10000 * N / 10000` pattern is the canonical
-///   wedge case). Requires `z3` on `PATH` when running `cargo kani --tests`.
-///
-/// `dsl_field_type` is the DSL-level type string from the spec
-/// (`U64`, `U128`, `I128`, `U8`, etc.), pre-`map_type`.
-fn pick_arith_solver(dsl_field_type: &str, rhs_is_arithmetic: bool) -> &'static str {
-    if !rhs_is_arithmetic {
-        return "cadical";
-    }
-    let is_wide = matches!(dsl_field_type, "U64" | "U128" | "I128");
-    if is_wide {
-        "bin = \"z3\""
-    } else {
-        "minisat"
-    }
-}
-
-/// Pick a solver for an effect RHS, chasing through the handler's `let`
-/// bindings. The canonical heavy-arith pattern hides behind a binding:
-///
-///     let total_fee = amount * 125 / 10000
-///     let net = amount - total_fee
-///     effect { pool += net, fees += total_fee }
-///
-/// Both effect RHSs are bare identifiers. A purely syntactic
-/// `pick_kani_solver("U64", "net")` returns cadical and wedges CBMC on
-/// a u64 mul/div symbolic exploration. Transitively resolving through the
-/// bindings exposes `total_fee`'s mul/div and routes the wide-LHS fields
-/// to z3.
-pub fn pick_kani_solver_for_effect(
-    dsl_field_type: &str,
-    rhs: &str,
-    op: &ParsedHandler,
-) -> &'static str {
-    // Fixed-point taint propagation: a binding is "arith-tainted" when its
-    // (transitive) RHS contains `*` or `/`. Bounded by the binding count.
-    let mut tainted: std::collections::HashSet<&str> = std::collections::HashSet::new();
-    for (name, _, bound_rhs) in &op.let_bindings {
-        if bound_rhs.contains('*') || bound_rhs.contains('/') {
-            tainted.insert(name.as_str());
-        }
-    }
-    for _ in 0..op.let_bindings.len() {
-        let mut changed = false;
-        for (name, _, bound_rhs) in &op.let_bindings {
-            if tainted.contains(name.as_str()) {
-                continue;
-            }
-            if tainted.iter().any(|t| contains_whole_word(bound_rhs, t)) {
-                tainted.insert(name.as_str());
-                changed = true;
-            }
-        }
-        if !changed {
-            break;
-        }
-    }
-
-    let rhs_is_arith = rhs.contains('*')
-        || rhs.contains('/')
-        || tainted.iter().any(|t| contains_whole_word(rhs, t));
-    pick_arith_solver(dsl_field_type, rhs_is_arith)
-}
-
-/// True if `hay` contains `needle` as a whole word (not a substring of a
-/// longer identifier). `net` in `amount - net` matches; `net` in `network`
-/// does not.
-fn contains_whole_word(hay: &str, needle: &str) -> bool {
-    if needle.is_empty() {
-        return false;
-    }
-    let bytes = hay.as_bytes();
-    let n = needle.as_bytes();
-    let mut i = 0;
-    while i + n.len() <= bytes.len() {
-        if &bytes[i..i + n.len()] == n {
-            let before_ok = i == 0 || !is_ident_byte(bytes[i - 1]);
-            let after_ok = i + n.len() == bytes.len() || !is_ident_byte(bytes[i + n.len()]);
-            if before_ok && after_ok {
-                return true;
-            }
-        }
-        i += 1;
-    }
-    false
-}
-
-fn is_ident_byte(b: u8) -> bool {
-    b.is_ascii_alphanumeric() || b == b'_'
-}
-
-/// Resolve an effect value to a Rust expression: handler param name,
-/// declared constant, `let … = call` binding, state field (rebound to
-/// `<state_binder>X` when provided), or pass-through literal.
-///
-/// State fields need a binder because upstream effect-RHS rendering already
-/// stripped the `state.` prefix (chumsky_adapter::render_effect) and each
-/// target binds state differently (proptest `s`, Anchor `self.<acct>`, …);
-/// a bare field name would be E0425 at compile time. Pass `None` for
-/// pass-through (bare identifier).
-pub fn resolve_value(
-    value: &str,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    state_binder: Option<&str>,
-) -> String {
-    if op.takes_params.iter().any(|(n, _)| n == value) {
-        value.to_string()
-    } else if let Some((_, const_val)) = spec.constants.iter().find(|(n, _)| n == value) {
-        const_val.clone()
-    } else if op
-        .calls
-        .iter()
-        .any(|c| c.result_binding.as_deref() == Some(value))
-    {
-        // `let <name> = call …` binding is in scope for subsequent
-        // effects / requires; render as the bare let-bound local.
-        value.to_string()
-    } else if let Some(binder) = state_binder {
-        if is_state_field(value, spec) {
-            format!("{}{}", binder, value)
-        } else {
-            value.to_string()
-        }
-    } else {
-        value.to_string()
-    }
-}
-
-pub fn resolve_value_with_account_env(
-    value: &str,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    state_binder: Option<&str>,
-    account_binder: Option<&str>,
-) -> String {
-    if let Some(binder) = account_binder {
-        let rewritten = rewrite_account_pubkey_refs(value, &op.accounts, binder);
-        if rewritten != value {
-            return rewritten;
-        }
-    }
-    resolve_value(value, op, spec, state_binder)
-}
-
-/// True when the bare identifier names a state field in the flat
-/// `state_fields` list or any `account_types[*].fields` (multi-account).
-fn is_state_field(name: &str, spec: &ParsedSpec) -> bool {
-    if spec.state_fields.iter().any(|(n, _)| n == name) {
-        return true;
-    }
-    for acct in &spec.account_types {
-        if acct.fields.iter().any(|(n, _)| n == name) {
-            return true;
-        }
-    }
-    false
-}
-
-// ============================================================================
-// Shared helpers — used by kani, proptest, unit_test, integration generators
-// ============================================================================
-
-/// Resolve state fields for the spec, handling multi-account layout.
-/// Returns the fields for the primary account type.
-pub fn resolve_state_fields(spec: &ParsedSpec) -> &[(String, String)] {
-    if spec.account_types.len() > 1 {
-        &spec.account_types[0].fields
-    } else {
-        &spec.state_fields
-    }
-}
-
-/// Every declared state field flows through here (the "mutable" naming is
-/// historical; Pubkey fields are first-class since they lower to
-/// `[u8; 32]`).
-pub fn mutable_fields(fields: &[(String, String)]) -> Vec<&(String, String)> {
-    fields.iter().collect()
-}
-
-/// True when the named field's declared type is `Pubkey`. Looks in the
-/// handler's target account first (multi-account specs), then falls back
-/// to global `state_fields`. Returns `false` if the field isn't found
-/// (callers default to "not Pubkey" — emit normally) so unknown fields
-/// surface as compile errors at the right line, not as silent skips.
-pub fn field_type_is_pubkey(field: &str, op: &ParsedHandler, spec: &ParsedSpec) -> bool {
-    // Variant-prefixed paths (`Active.owner`) resolve against the variant's
-    // payload first; fall through to the flat schema otherwise.
-    if let Some(dot) = field.find('.') {
-        let head = &field[..dot];
-        let rest = &field[dot + 1..];
-        let nested_base = effect_target_base(rest);
-        for at in &spec.account_types {
-            if let Some(variant) = at.variants.iter().find(|v| v.name == head) {
-                if let Some((_, t)) = variant.fields.iter().find(|(n, _)| n == nested_base) {
-                    return t == "Pubkey";
-                }
-            }
-        }
-    }
-    let base = effect_target_base(field);
-    if let Some(ref acct_name) = op.on_account {
-        if let Some(acct) = spec.account_types.iter().find(|a| a.name == *acct_name) {
-            if let Some((_, t)) = acct.fields.iter().find(|(n, _)| n == base) {
-                return t == "Pubkey";
-            }
-        }
-    }
-    spec.state_fields
-        .iter()
-        .find(|(n, _)| n == base)
-        .map(|(_, t)| t == "Pubkey")
-        .unwrap_or(false)
-}
-
-/// The base field name an effect targets. `accounts[i].active` → `accounts`;
-/// `foo.bar` → `foo`; `plain` → `plain`. Used by `check_effect_targets` to
-/// look up the target in the declared state schema.
-pub fn effect_target_base(path: &str) -> &str {
-    let path = path.trim();
-    let end = path.find(['[', '.']).unwrap_or(path.len());
-    &path[..end]
-}
-
-/// Strip a leading `<Variant>.` prefix when the root names a multi-variant
-/// ADT variant; unchanged otherwise. Harness `State` models carry fields in
-/// union form, so `Active.balance := …` must lower to `s.balance = …`.
-/// Owned return so callers can pass through `&str`-only APIs.
-pub fn strip_variant_prefix_for_flat_state(path: &str, spec: &ParsedSpec) -> String {
-    if let Some(dot) = path.find('.') {
-        let head = &path[..dot];
-        let is_variant = spec
-            .account_types
-            .iter()
-            .any(|a| a.variants.iter().any(|v| v.name == head));
-        if is_variant {
-            return path[dot + 1..].to_string();
-        }
-    }
-    path.to_string()
-}
-
-/// Project an effect-shaped MIR `Stmt` back onto the `(field, op_kind,
-/// value)` triple the string templates below consume — the #66 adaptor
-/// that makes `Stmt` the iteration source for the Kani/proptest transition
-/// bodies while keeping output byte-identical (`lower_body` preserves spec
-/// order, `Path` round-trips the dotted field, `Expr::from_raw` carries
-/// the RHS verbatim in `.rust`).
-///
-/// Returns `None` for every non-effect variant — each with the reason it
-/// renders as *nothing* in the pure spec-model transition. The match is
-/// exhaustive by discipline (no `_` arm; see the `Stmt` enum doc): a new
-/// `Stmt` variant is a compile error here, forcing an explicit decision
-/// for the Kani/proptest backends.
-pub fn stmt_effect_triple(stmt: &crate::mir::Stmt) -> Option<(String, &'static str, &str)> {
-    use crate::mir::Stmt;
-    match stmt {
-        Stmt::Assign { path, rhs } => Some((path.segments.join("."), "set", rhs.rust.as_str())),
-        Stmt::CheckedAdd { path, delta, .. } => {
-            // The lowered per-site error name is Lean/scaffold surface;
-            // the pure model signals overflow via `return false`.
-            Some((path.segments.join("."), "add", delta.rust.as_str()))
-        }
-        Stmt::CheckedSub { path, delta, .. } => {
-            Some((path.segments.join("."), "sub", delta.rust.as_str()))
-        }
-        Stmt::SatAdd { path, delta } => {
-            Some((path.segments.join("."), "add_sat", delta.rust.as_str()))
-        }
-        Stmt::SatSub { path, delta } => {
-            Some((path.segments.join("."), "sub_sat", delta.rust.as_str()))
-        }
-        Stmt::WrapAdd { path, delta } => {
-            Some((path.segments.join("."), "add_wrap", delta.rust.as_str()))
-        }
-        Stmt::WrapSub { path, delta } => {
-            Some((path.segments.join("."), "sub_wrap", delta.rust.as_str()))
-        }
-        // Guard surface: `requires X else Err` folds into the guard
-        // conjunction via `collect_full_guard*`, not the body.
-        Stmt::RequireOrAbort { .. } => None,
-        // CPI surface: no state mutation in the pure model; the Kani
-        // CPI-fact harnesses read the call sites separately.
-        Stmt::TokenTransfer { .. } => None,
-        Stmt::Cpi { .. } => None,
-        // Lifecycle surface: the transition body drives variant changes
-        // through the pre/post-status writes, not a promote statement.
-        Stmt::VariantPromote { .. } => None,
-        // Branch arms are rendered by the per-arm match path; walking
-        // them here would double-emit.
-        Stmt::Branch { .. } => None,
-        // Abort clauses are harnessed from the `aborts_if` predicate
-        // surface; in the body they carry no state mutation.
-        Stmt::Abort(_) => None,
-        // Events: auxiliary, no state mutation in the pure model.
-        Stmt::Emit { .. } => None,
-    }
-}
-
-/// All effect triples of a lowered handler body, in spec order. The
-/// shared iteration source for the Kani/proptest transition emitters,
-/// conformance harnesses, and overflow filters. Top-level only —
-/// `Stmt::Branch` arms are NOT descended (per-arm rendering owns
-/// those); use [`block_effect_triples_deep`] for analyses that must
-/// see conditionally-applied effects.
-pub fn block_effect_triples(body: &crate::mir::Block) -> Vec<(String, &'static str, &str)> {
-    body.stmts.iter().filter_map(stmt_effect_triple).collect()
-}
-
-/// Like [`block_effect_triples`] but descends into `Stmt::Branch`
-/// arms/default. For *may-this-effect-fire* analyses (overflow filters
-/// and tests) where a conditionally-applied effect counts.
-pub fn block_effect_triples_deep(body: &crate::mir::Block) -> Vec<(String, &'static str, &str)> {
-    fn walk<'a>(stmts: &'a [crate::mir::Stmt], out: &mut Vec<(String, &'static str, &'a str)>) {
-        for s in stmts {
-            if let Some(t) = stmt_effect_triple(s) {
-                out.push(t);
-            }
-            if let crate::mir::Stmt::Branch { arms, default, .. } = s {
-                for a in arms {
-                    walk(&a.block.stmts, out);
-                }
-                if let Some(d) = default {
-                    walk(&d.stmts, out);
-                }
-            }
-        }
-    }
-    let mut out = Vec::new();
-    walk(&body.stmts, &mut out);
-    out
-}
-
-/// Render a single `(field, op_kind, value)` triple into Rust at the given
-/// indent. The helper writes the trailing newline; the caller controls
-/// where the statement sits relative to its surrounding block.
-#[allow(clippy::too_many_arguments)]
-pub fn emit_one_effect(
-    out: &mut String,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    wrapping: bool,
-    field: &str,
-    op_kind: &str,
-    value: &str,
-    indent: &str,
-) {
-    emit_one_effect_inner(out, op, spec, wrapping, field, op_kind, value, indent, None);
-}
-
-#[allow(clippy::too_many_arguments)]
-fn emit_one_effect_inner(
-    out: &mut String,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    wrapping: bool,
-    field: &str,
-    op_kind: &str,
-    value: &str,
-    indent: &str,
-    account_binder: Option<&str>,
-) {
-    // Harnesses run against a flat `State` (union-of-variant-fields view):
-    // strip the variant prefix so `Variant.field := …` emits `s.field = …`;
-    // the variant itself is tracked via `s.status`.
-    let field_owned = strip_variant_prefix_for_flat_state(field, spec);
-    let field = field_owned.as_str();
-    // Body binds state as `s` — pass that binder so a bare state-field RHS
-    // renders as `s.<field>`.
-    let rust_value = resolve_value_with_account_env(value, op, spec, Some("s."), account_binder);
-    match op_kind {
-        "set" => {
-            out.push_str(&format!("{indent}s.{field} = {rust_value};\n"));
-        }
-        "add" => {
-            if wrapping {
-                out.push_str(&format!(
-                    "{indent}s.{field} = s.{field}.wrapping_add({rust_value});\n"
-                ));
-            } else {
-                out.push_str(&format!(
-                    "{indent}match s.{field}.checked_add({rust_value}) {{\n\
-                     {indent}    Some(__v) => s.{field} = __v,\n\
-                     {indent}    None => return false,\n\
-                     {indent}}}\n"
-                ));
-            }
-        }
-        "add_sat" => {
-            out.push_str(&format!(
-                "{indent}s.{field} = s.{field}.saturating_add({rust_value});\n"
-            ));
-        }
-        "add_wrap" => {
-            out.push_str(&format!(
-                "{indent}s.{field} = s.{field}.wrapping_add({rust_value});\n"
-            ));
-        }
-        "sub" => {
-            if wrapping {
-                out.push_str(&format!(
-                    "{indent}s.{field} = s.{field}.wrapping_sub({rust_value});\n"
-                ));
-            } else {
-                out.push_str(&format!(
-                    "{indent}match s.{field}.checked_sub({rust_value}) {{\n\
-                     {indent}    Some(__v) => s.{field} = __v,\n\
-                     {indent}    None => return false,\n\
-                     {indent}}}\n"
-                ));
-            }
-        }
-        "sub_sat" => {
-            out.push_str(&format!(
-                "{indent}s.{field} = s.{field}.saturating_sub({rust_value});\n"
-            ));
-        }
-        "sub_wrap" => {
-            out.push_str(&format!(
-                "{indent}s.{field} = s.{field}.wrapping_sub({rust_value});\n"
-            ));
-        }
-        _ => {
-            out.push_str(&format!(
-                "{indent}// unknown effect: {field} {op_kind} {value}\n"
-            ));
-        }
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub fn emit_one_effect_with_account_env(
-    out: &mut String,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    wrapping: bool,
-    field: &str,
-    op_kind: &str,
-    value: &str,
-    indent: &str,
-    account_binder: &str,
-) {
-    emit_one_effect_inner(
-        out,
-        op,
-        spec,
-        wrapping,
-        field,
-        op_kind,
-        value,
-        indent,
-        Some(account_binder),
-    );
-}
-
-/// Verify every effect-target field is declared somewhere in the state
-/// schema (`state_fields`, per-account fields, or a sum-variant payload).
-/// Errors name the handler and field — catching this at codegen time beats
-/// a `cargo check` error 1000 lines into the generated harness.
-pub fn check_effect_targets(spec: &ParsedSpec) -> anyhow::Result<()> {
-    use std::collections::HashSet;
-
-    // Collect every declared field name from every place fields can live.
-    let mut declared: HashSet<&str> = HashSet::new();
-    for (n, _) in &spec.state_fields {
-        declared.insert(n.as_str());
-    }
-    for acct in &spec.account_types {
-        for (n, _) in &acct.fields {
-            declared.insert(n.as_str());
-        }
-    }
-    for rec in &spec.records {
-        for (n, _) in &rec.fields {
-            declared.insert(n.as_str());
-        }
-    }
-    for sum in &spec.sum_types {
-        for variant in &sum.variants {
-            for (n, _) in &variant.fields {
-                declared.insert(n.as_str());
-            }
-        }
-    }
-
-    // Variant-prefixed targets (`Active.balance`) are legal: index variant
-    // fields so the check re-targets at the field beneath the prefix
-    // instead of false-positive-bailing on the variant name.
-    let mut variant_fields: std::collections::HashMap<&str, HashSet<&str>> =
-        std::collections::HashMap::new();
-    for acct in &spec.account_types {
-        for variant in &acct.variants {
-            let entry = variant_fields.entry(variant.name.as_str()).or_default();
-            for (n, _) in &variant.fields {
-                entry.insert(n.as_str());
-            }
-        }
-    }
-
-    for handler in &spec.handlers {
-        for (field, _kind, _value) in &handler.effects {
-            let base = effect_target_base(field);
-            // Variant-prefixed: the root is a variant name, so check
-            // the field beneath it against that variant's payload.
-            if let Some(variant_payload) = variant_fields.get(base) {
-                let after = field.trim_start_matches(base).trim_start_matches('.');
-                let nested_base = effect_target_base(after);
-                if !nested_base.is_empty()
-                    && !variant_payload.contains(nested_base)
-                    && !declared.contains(nested_base)
-                {
-                    anyhow::bail!(
-                        "handler `{}` writes effect target `{}` but `{}` is not declared in variant `{}`'s payload — add it to the variant or rename the effect",
-                        handler.name,
-                        field,
-                        nested_base,
-                        base,
-                    );
-                }
-                continue;
-            }
-            if !declared.contains(base) {
-                // `state := .Variant { … }` desugars to per-field effects
-                // at the adapter, but non-RecordLit / unit-variant shapes
-                // can survive here as a single bare-`state` effect. Accept
-                // `state` whenever the spec has a multi-variant ADT —
-                // downstream either handles it (RecordLit) or bails to a
-                // `todo!()`.
-                if base == "state" && spec.account_types.iter().any(|a| !a.variants.is_empty()) {
-                    continue;
-                }
-                anyhow::bail!(
-                    "handler `{}` writes effect target `{}` but `{}` is not declared in any state, account, record, or sum-variant payload — add it to the state declaration or remove the effect",
-                    handler.name,
-                    field,
-                    base,
-                );
-            }
-        }
-    }
-    Ok(())
-}
-
-/// Collect guard_str + requires clauses as a single Rust expression; None
-/// if no guards. Skips `requires` bodies referencing
-/// `<handler-account>.pubkey` — the harness `State` model doesn't carry
-/// handler accounts, so they'd be compile errors. The runtime-side check
-/// still emits in the real handler; only the property-test projection
-/// drops it (same shape as the lean_gen drop).
-pub fn collect_full_guard(op: &ParsedHandler, wrapping: bool) -> Option<String> {
-    collect_full_guard_with_account_env(op, wrapping, None)
-}
-
-pub fn collect_full_guard_with_account_env(
-    op: &ParsedHandler,
-    wrapping: bool,
-    account_binder: Option<&str>,
-) -> Option<String> {
-    let mut parts = Vec::new();
-    if let Some(ref guard) = op.guard_str {
-        let translated = translate_guard_to_rust(guard, wrapping);
-        let translated = account_binder
-            .map(|binder| rewrite_account_pubkey_refs(&translated, &op.accounts, binder))
-            .unwrap_or(translated);
-        parts.push(format!("({})", translated));
-    }
-    for req in &op.requires {
-        if account_binder.is_none() && mentions_handler_account_pubkey(&req.rust_expr, &op.accounts)
-        {
-            continue;
-        }
-        let translated = translate_guard_to_rust(&req.rust_expr, wrapping);
-        let translated = account_binder
-            .map(|binder| rewrite_account_pubkey_refs(&translated, &op.accounts, binder))
-            .unwrap_or(translated);
-        parts.push(format!("({})", translated));
-    }
-    if parts.is_empty() {
-        None
-    } else {
-        Some(parts.join(" && "))
-    }
-}
-
-pub fn collect_guard_terms_with_account_env(
-    op: &ParsedHandler,
-    wrapping: bool,
-    account_binder: Option<&str>,
-) -> Vec<GuardTerm> {
-    let mut terms = Vec::new();
-    if let Some(ref guard) = op.guard_str {
-        let translated = translate_guard_to_rust(guard, wrapping);
-        let translated = account_binder
-            .map(|binder| rewrite_account_pubkey_refs(&translated, &op.accounts, binder))
-            .unwrap_or(translated);
-        push_split_guard_terms(&mut terms, GuardTermSource::Guard, &translated);
-    }
-    for req in &op.requires {
-        if account_binder.is_none() && mentions_handler_account_pubkey(&req.rust_expr, &op.accounts)
-        {
-            continue;
-        }
-        let translated = translate_guard_to_rust(&req.rust_expr, wrapping);
-        let translated = account_binder
-            .map(|binder| rewrite_account_pubkey_refs(&translated, &op.accounts, binder))
-            .unwrap_or(translated);
-        push_split_guard_terms(
-            &mut terms,
-            GuardTermSource::Requires {
-                error_name: req.error_name.clone(),
-            },
-            &translated,
-        );
-    }
-    terms
-}
-
-fn push_split_guard_terms(terms: &mut Vec<GuardTerm>, source: GuardTermSource, expr: &str) {
-    for term in split_top_level_and(expr) {
-        terms.push(GuardTerm {
-            source: source.clone(),
-            rust_expr: term,
-        });
-    }
-}
-
-pub fn split_top_level_and(expr: &str) -> Vec<String> {
-    let bytes = expr.as_bytes();
-    let mut parts = Vec::new();
-    let mut start = 0usize;
-    let mut paren_depth = 0usize;
-    let mut bracket_depth = 0usize;
-    let mut brace_depth = 0usize;
-    let mut in_string = false;
-    let mut escaped = false;
-    let mut i = 0usize;
-
-    while i < bytes.len() {
-        let b = bytes[i];
-        if in_string {
-            if escaped {
-                escaped = false;
-            } else if b == b'\\' {
-                escaped = true;
-            } else if b == b'"' {
-                in_string = false;
-            }
-            i += 1;
-            continue;
-        }
-
-        match b {
-            b'"' => in_string = true,
-            b'(' => paren_depth += 1,
-            b')' => paren_depth = paren_depth.saturating_sub(1),
-            b'[' => bracket_depth += 1,
-            b']' => bracket_depth = bracket_depth.saturating_sub(1),
-            b'{' => brace_depth += 1,
-            b'}' => brace_depth = brace_depth.saturating_sub(1),
-            b'&' if i + 1 < bytes.len()
-                && bytes[i + 1] == b'&'
-                && paren_depth == 0
-                && bracket_depth == 0
-                && brace_depth == 0 =>
-            {
-                let part = expr[start..i].trim();
-                if !part.is_empty() {
-                    parts.push(part.to_string());
-                }
-                i += 2;
-                start = i;
-                continue;
-            }
-            _ => {}
-        }
-        i += 1;
-    }
-
-    let tail = expr[start..].trim();
-    if !tail.is_empty() {
-        parts.push(tail.to_string());
-    }
-    parts
-}
-
-/// True when `expr` mentions `<handler_account>.pubkey` (or `.key()`) —
-/// used to suppress such `requires` from property-test guard collection.
-fn mentions_handler_account_pubkey(
-    expr: &str,
-    accounts: &[crate::check::ParsedHandlerAccount],
-) -> bool {
-    accounts.iter().any(|a| {
-        let needle_pubkey = format!("{}.pubkey", a.name);
-        let needle_key = format!("{}.key()", a.name);
-        expr.contains(&needle_pubkey) || expr.contains(&needle_key)
-    })
-}
-
-// ============================================================================
-// Shared emitters
-// ============================================================================
-
-/// Emit `let <name>: <T> = <source>;` for each `abstract <name> : <T>`
-/// binder. `source` is the per-backend symbolic-input expression
-/// (`kani::any()`, `todo!("…")`, …). Call after takes_params emission so
-/// the binders are in scope for the following assume/prop_assume reads.
-pub fn emit_abstract_binders(
-    out: &mut String,
-    handler: &crate::check::ParsedHandler,
-    indent: &str,
-    source: &str,
-    map_ty: impl Fn(&str) -> anyhow::Result<String>,
-) -> anyhow::Result<()> {
-    for (name, ty_str) in &handler.abstract_binders {
-        let ty = map_ty(ty_str)?;
-        out.push_str(&format!("{}let {}: {} = {};\n", indent, name, ty, source));
-    }
-    Ok(())
-}
-
-/// Emit `let mut s = State { ... };` with every mutable field bound to
-/// `kani::any()`. When the per-account lifecycle has ≥2 states, the
-/// synthetic `status` field is also `kani::any()` so callers can layer
-/// `kani::assume(s.status == Status::<X>)` on top.
-pub fn emit_state_init_symbolic(
-    out: &mut String,
-    mutable_fields: &[&(String, String)],
-    lifecycle_states: &[String],
-) {
-    out.push_str("    let mut s = State {\n");
-    for (fname, _) in mutable_fields {
-        out.push_str(&format!("        {}: kani::any(),\n", fname));
-    }
-    if lifecycle_states.len() >= 2 {
-        out.push_str("        status: kani::any(),\n");
-    }
-    out.push_str("    };\n");
-}
-
-/// Emit `let mut s = State { ... };` zeroed, with `status` set to the
-/// initial lifecycle state — the canonical pre-state for init-handler
-/// harnesses. Type-aware defaults come from the shared DSL type surface.
-pub fn emit_state_init_zeroed(
-    out: &mut String,
-    mutable_fields: &[&(String, String)],
-    lifecycle_states: &[String],
-    spec: &crate::check::ParsedSpec,
-) {
-    out.push_str("    let mut s = State {\n");
-    for (fname, ftype) in mutable_fields {
-        if let Some(default) = spec.default_value_for_type(ftype) {
-            out.push_str(&format!("        {}: {},\n", fname, default));
-        }
-    }
-    if let Some(initial) = lifecycle_states.first() {
-        if lifecycle_states.len() >= 2 {
-            out.push_str(&format!("        status: Status::{},\n", initial));
-        }
-    }
-    out.push_str("    };\n");
-}
-
-/// Append `kani::assume(s.status == Status::<pre>);` when the handler has a
-/// pre-status declaration AND this section has a lifecycle; no-op otherwise.
-/// Without this, guard-rejection / abort harnesses can pass for the wrong
-/// reason — the handler rejects on a mismatched symbolic status, not
-/// because the requires/guard fired.
-pub fn emit_pre_status_assume(
-    out: &mut String,
-    op: &crate::check::ParsedHandler,
-    lifecycle_states: &[String],
-) {
-    if lifecycle_states.len() < 2 {
-        return;
-    }
-    if let Some(ref pre) = op.pre_status {
-        out.push_str(&format!("    kani::assume(s.status == Status::{});\n", pre));
-    }
-}
-
-pub fn emit_constants(out: &mut String, constants: &[(String, String)]) {
-    for (name, value) in constants {
-        let upper = name.to_uppercase();
-        let const_type = infer_const_type(value);
-        out.push_str(&format!("const {}: {} = {};\n", upper, const_type, value));
-    }
-    if !constants.is_empty() {
-        out.push('\n');
-    }
-}
-
-/// Emit struct declarations for user-defined record types. Called before
-/// `emit_state_struct` so records are in scope when State references them.
-/// `derives` is the per-backend `#[derive(...)]` list. Empty records are
-/// skipped.
-pub fn emit_record_structs(
-    out: &mut String,
-    spec: &crate::check::ParsedSpec,
-    derives: &str,
-    map_type_fn: impl Fn(&str) -> anyhow::Result<String>,
-) -> anyhow::Result<()> {
-    for rec in &spec.records {
-        if rec.fields.is_empty() {
-            continue;
-        }
-        // Flat `state { … }` forms produce a record literally named
-        // `State`; the state-machine `struct State` (lifecycle + ghost
-        // fields) is emitted separately, so skip to avoid a duplicate.
-        if rec.name == "State" {
-            continue;
-        }
-        out.push_str(&format!("#[derive({})]\n", derives));
-        out.push_str(&format!("struct {} {{\n", rec.name));
-        for (fname, ftype) in &rec.fields {
-            out.push_str(&format!("    {}: {},\n", fname, map_type_fn(ftype)?));
-        }
-        out.push_str("}\n\n");
-    }
-    Ok(())
-}
-
-/// Emit enums for sum-types whose variants are ALL unit (`type Error |
-/// NotAdmin | …` → `enum Error { NotAdmin, … }`). Payload-carrying sums
-/// (`type State | Active of { … }`) are skipped — codegen flattens those
-/// into a `struct State`, and an `enum State` would collide.
-pub fn emit_unit_enum_sums(
-    out: &mut String,
-    spec: &crate::check::ParsedSpec,
-    derives: &str,
-) -> anyhow::Result<()> {
-    for sum in &spec.sum_types {
-        let all_unit = sum.variants.iter().all(|v| v.fields.is_empty());
-        if !all_unit || sum.variants.is_empty() {
-            continue;
-        }
-        out.push_str(&format!("#[derive({})]\n", derives));
-        out.push_str(&format!("enum {} {{\n", sum.name));
-        for variant in &sum.variants {
-            out.push_str(&format!("    {},\n", variant.name));
-        }
-        out.push_str("}\n\n");
-    }
-    Ok(())
-}
-
-/// True when the spec declares a multi-state lifecycle the harness layer
-/// should model as a `Status` enum + `status` field; single-state / no
-/// lifecycle needs no discriminator.
-pub fn has_lifecycle(spec: &crate::check::ParsedSpec) -> bool {
-    spec.lifecycle_states.len() >= 2
-}
-
-/// Emit the synthetic `Status` enum from a per-account or per-spec
-/// lifecycle slice; no-op below two states. Synthetic: derived from the
-/// State sum-type's variants, not user-declared — without a status field,
-/// lifecycle-only handlers have nothing to write and every harness against
-/// them is vacuous. Multi-ADT codegen must pass `acct.lifecycle` so each
-/// `mod <acct>` gets its own variants, not the spec-level ones.
-pub fn emit_lifecycle_status_enum_from(
-    out: &mut String,
-    lifecycle_states: &[String],
-    derives: &str,
-) {
-    if lifecycle_states.len() < 2 {
-        return;
-    }
-    out.push_str(&format!("#[derive({})]\n", derives));
-    out.push_str("enum Status {\n");
-    for state in lifecycle_states {
-        out.push_str(&format!("    {},\n", state));
-    }
-    out.push_str("}\n\n");
-}
-
-/// Emit a State struct with configurable derives. `map_type_fn` errors on
-/// unrecognized DSL types so codegen fails loudly. `has_lifecycle` gates
-/// the `status: Status` field — multi-ADT codegen threads the per-account
-/// lifecycle, not the spec-level one. Callers must have already emitted
-/// the `Status` enum via `emit_lifecycle_status_enum_from`.
-pub fn emit_state_struct_with_lifecycle(
-    out: &mut String,
-    fields: &[&(String, String)],
-    derives: &str,
-    map_type_fn: impl Fn(&str) -> anyhow::Result<String>,
-    has_lifecycle: bool,
-) -> anyhow::Result<()> {
-    out.push_str(&format!("#[derive({})]\n", derives));
-    out.push_str("struct State {\n");
-    for (fname, ftype) in fields {
-        out.push_str(&format!("    {}: {},\n", fname, map_type_fn(ftype)?));
-    }
-    if has_lifecycle && !fields.iter().any(|(n, _)| n == "status") {
-        out.push_str("    status: Status,\n");
-    }
-    out.push_str("}\n\n");
-    Ok(())
-}
-
-/// Emit `fn {inv_name}(s: &State) -> bool { <rust_expr> }` per invariant
-/// with a Rust body. Description-only invariants and unsupported
-/// quantifier bodies are skipped silently; callers pre-filter to the
-/// invariants relevant for the current account section / state shape.
-pub fn emit_invariant_predicates(out: &mut String, invariants: &[&crate::check::ParsedInvariant]) {
-    for inv in invariants {
-        let Some(rust_expr) = inv.rust_expr.as_deref() else {
-            continue;
-        };
-        if crate::check::rust_expr_is_unsupported(rust_expr) {
-            continue;
-        }
-        let doc_body = inv
-            .lean_expr
-            .as_deref()
-            .map(|le| format!(" — {}", le))
-            .unwrap_or_default();
-        out.push_str(&format!("/// Invariant: {}{}\n", inv.name, doc_body));
-        out.push_str(&format!("fn {}(s: &State) -> bool {{\n", inv.name));
-        out.push_str(&format!("    {}\n", rust_expr));
-        out.push_str("}\n\n");
-    }
-}
-
-/// Emit property predicate functions. `map_type_fn` lets the per-slot
-/// `<prop>_at(s, <binder>)` predicate render a target-specific binder type
-/// (Quasar Pod vs native Rust differ for non-primitive binders).
-///
-/// Emission shape:
-///   - Always `fn <prop>(s: &State) -> bool` — the real expression, or
-///     `true` when the body has a quantifier (the harness drives the check
-///     via `<prop>_at` instead).
-///   - When `prop.per_slot` is Some, also `fn <prop>_at(s: &State,
-///     <binder>: <ty>) -> bool` — the `forall` inner expression with the
-///     binder free; harnesses bind it symbolically for a non-vacuous check.
-pub fn emit_property_predicates_with(
-    out: &mut String,
-    properties: &[ParsedProperty],
-    wrapping: bool,
-    map_type_fn: impl Fn(&str) -> anyhow::Result<String>,
-) {
-    for prop in properties {
-        // Prefer the AST-rendered Rust form (handles implies/forall, embeds
-        // the `QEDGEN_UNSUPPORTED_QUANTIFIER` marker); fall back to
-        // translating the Lean form for callers without an AST.
-        let rendered = prop
-            .rust_expression
-            .as_deref()
-            .map(|r| r.to_string())
-            .or_else(|| {
-                prop.expression
-                    .as_deref()
-                    .map(|e| translate_property_to_rust(e, wrapping))
-            });
-        let Some(rust_expr) = rendered else { continue };
-        let doc = prop.expression.as_deref().unwrap_or("");
-        out.push_str(&format!("/// {}: {}\n", prop.name, doc));
-        // Binary properties (body contains `old(...)`) take `(pre, post)`;
-        // the adapter renders `state.x` → `post.x`, `old(state.x)` →
-        // `pre.x`. Kani's preservation harness dispatches assertion arity
-        // on `prop.class`.
-        let is_binary = prop.class == crate::check::PropertyClass::Binary;
-        let sig = if is_binary {
-            format!("fn {}(pre: &State, post: &State) -> bool", prop.name)
-        } else {
-            format!("fn {}(s: &State) -> bool", prop.name)
-        };
-        // Stubs underscore the params so the body `true` doesn't trip
-        // `unused_variables`.
-        let stub_sig = if is_binary {
-            format!("fn {}(_pre: &State, _post: &State) -> bool", prop.name)
-        } else {
-            format!("fn {}(_s: &State) -> bool", prop.name)
-        };
-        if crate::check::rust_expr_is_unsupported(&rust_expr) {
-            // Quantifier body: emit a `true` stub; the harness preamble
-            // skips calling into these predicates.
-            out.push_str(&format!("{} {{\n", stub_sig));
-            out.push_str(&format!(
-                "    // {} — property uses a quantifier; lower at the harness level.\n",
-                rust_expr.trim()
-            ));
-            out.push_str("    true\n");
-            out.push_str("}\n\n");
-        } else {
-            out.push_str(&format!("{} {{\n", sig));
-            out.push_str(&format!("    {}\n", rust_expr));
-            out.push_str("}\n\n");
-        }
-        // Per-slot predicate: the adapter populates `per_slot` for
-        // mechanically-lowerable `forall <binder> : <ty>, body` properties;
-        // harnesses bind `<binder>` symbolically and call `<prop>_at`.
-        if let Some(slot) = &prop.per_slot {
-            let rust_ty =
-                map_type_fn(&slot.binder_type).unwrap_or_else(|_| slot.binder_type.clone());
-            out.push_str(&format!(
-                "/// {}: per-slot check at `{}: {}` (v2.20 forall lowering)\n",
-                prop.name, slot.binder_name, slot.binder_type
-            ));
-            out.push_str("#[allow(unused_variables)]\n");
-            out.push_str(&format!(
-                "fn {}_at(s: &State, {}: {}) -> bool {{\n",
-                prop.name, slot.binder_name, rust_ty
-            ));
-            out.push_str(&format!("    {}\n", slot.rust_body));
-            out.push_str("}\n\n");
-        }
-    }
-}
-
-/// Emit `hook after_store(<field>)` assertions, anchored right after the
-/// field's effect so they see the post-store state. A failed assertion
-/// panics, which proptest/Kani surface as a failure. On-chain codegen
-/// never uses this emitter, so hooks don't reach the program.
-fn emit_after_store_hooks(
-    out: &mut String,
-    hooks: &[crate::mir::HookMir],
-    field: &str,
-    indent: &str,
-) {
-    let base = effect_target_base(field);
-    for hook in hooks {
-        if let crate::mir::HookKind::AfterStore(f) = &hook.kind {
-            if f == base {
-                for a in &hook.asserts {
-                    out.push_str(&format!(
-                        "{}assert!({}, \"hook after_store({}) violated\");\n",
-                        indent, a.rust, base
-                    ));
-                }
-            }
-        }
-    }
-}
-
-/// Both transition emitters iterate the handler's lowered MIR body for
-/// effects (`stmt_effect_triple`; #66 — a new `Stmt` variant is a compile
-/// error at the adaptor). The guard / status / let-binding / ghost
-/// scaffold stays `ParsedHandler`-fed by design — predicate/account
-/// surface, same boundary as `codegen_mir`'s guards.
-pub fn emit_transition_fn(
-    out: &mut String,
-    mir: &crate::mir::Mir,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    wrapping: bool,
-    map_type_fn: impl Fn(&str) -> anyhow::Result<String>,
-) -> anyhow::Result<()> {
-    emit_transition_fn_inner(out, mir, op, spec, wrapping, None, false, map_type_fn)
-}
-
-pub fn emit_transition_fn_for_kani(
-    out: &mut String,
-    mir: &crate::mir::Mir,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    wrapping: bool,
-    map_type_fn: impl Fn(&str) -> anyhow::Result<String>,
-) -> anyhow::Result<()> {
-    let account_env =
-        handler_needs_account_env(op).then(|| handler_account_env_struct_name(&op.name));
-    emit_transition_fn_inner(
-        out,
-        mir,
-        op,
-        spec,
-        wrapping,
-        account_env.as_deref(),
-        true,
-        map_type_fn,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-fn emit_transition_fn_inner(
-    out: &mut String,
-    mir: &crate::mir::Mir,
-    op: &ParsedHandler,
-    spec: &ParsedSpec,
-    wrapping: bool,
-    account_env_struct: Option<&str>,
-    rewrite_pubkey_comparisons: bool,
-    map_type_fn: impl Fn(&str) -> anyhow::Result<String>,
-) -> anyhow::Result<()> {
-    let body = mir
-        .handler_block(&op.name)
-        .ok_or_else(|| anyhow::anyhow!("MIR has no handler `{}`", op.name))?;
-    if let Some(ref doc) = op.doc {
-        out.push_str(&format!("/// {}\n", doc.trim()));
-    }
-
-    let mut params = String::new();
-    if let Some(account_env_struct) = account_env_struct {
-        params.push_str(&format!(", accounts: &{}", account_env_struct));
-    }
-    params.push_str(
-        &op.takes_params
-            .iter()
-            .chain(op.abstract_binders.iter())
-            .map(|(n, t)| map_type_fn(t).map(|rt| format!(", {}: {}", n, rt)))
-            .collect::<anyhow::Result<Vec<_>>>()?
-            .concat(),
-    );
-    // Abstract binders ride alongside real handler params; callers pass a
-    // symbolic / arbitrary value for each.
-    out.push_str(&format!(
-        "fn {}(s: &mut State{}) -> bool {{\n",
-        op.name, params
-    ));
-
-    // Guard check (merges guard_str + requires clauses)
-    if let Some(guard_expr) =
-        collect_full_guard_with_account_env(op, wrapping, account_env_struct.map(|_| "accounts"))
-    {
-        if let Some(ref raw) = op.guard_str {
-            out.push_str(&format!("    // guard: {}\n", raw));
-        }
-
-        let guard_terms = collect_guard_terms_with_account_env(
-            op,
-            wrapping,
-            account_env_struct.map(|_| "accounts"),
-        );
-        if rewrite_pubkey_comparisons && guard_terms.len() > 8 {
-            for term in guard_terms {
-                let term_expr = rewrite_kani_pubkey_comparisons(&term.rust_expr, op, spec);
-                if let Some(negated) = negate_simple_top_level_comparison(&term_expr) {
-                    out.push_str(&format!("    if {} {{\n", negated));
-                } else {
-                    out.push_str(&format!("    if !({}) {{\n", term_expr));
-                }
-                out.push_str("        return false;\n");
-                out.push_str("    }\n");
-            }
-        } else {
-            let guard_expr = if rewrite_pubkey_comparisons {
-                rewrite_kani_pubkey_comparisons(&guard_expr, op, spec)
-            } else {
-                guard_expr
-            };
-            out.push_str(&format!("    if !({}) {{\n", guard_expr));
-            out.push_str("        return false;\n");
-            out.push_str("    }\n");
-        }
-    }
-
-    // Pre-status check — handlers declared `State.X -> State.Y` must reject
-    // when the current lifecycle state isn't `X`. Without this, lifecycle-
-    // only handlers (whose effects don't touch user fields) would have
-    // empty bodies and every cover/liveness harness against them would
-    // pass tautologically.
-    if has_lifecycle(spec) {
-        if let Some(ref pre) = op.pre_status {
-            out.push_str(&format!("    if s.status != Status::{} {{\n", pre));
-            out.push_str("        return false;\n");
-            out.push_str("    }\n");
-        }
-    }
-
-    // Spec-level `let` bindings emit BEFORE the effect block so effect
-    // RHSs can reference them.
-    for (binding_name, _lean_expr, rust_expr) in &op.let_bindings {
-        out.push_str(&format!("    let {} = {};\n", binding_name, rust_expr));
-    }
-
-    // Apply effects. Per-effect arithmetic semantics: `+=` → checked_add
-    // (short-circuit via `return false`, matching deployed
-    // `checked_add(..).ok_or(err)?`), `+=!` → saturating, `+=?` → wrapping
-    // (same tiers for `-=`). The `wrapping` flag forces default `+=`/`-=`
-    // to wrap (proptest full-state-space mode); explicit `+=!`/`+=?`
-    // always honor their declared semantics.
-    //
-    // Effects targeting `Pubkey` fields are skipped when there's no
-    // account env: accounts aren't carried into the pure model, and pubkey
-    // identity is validated by the accounts struct at handler entry.
-    //
-    // `match` inside `effect { … }` lowers to `Stmt::Branch` (suppressing
-    // the flat union `op.effects` still carries for back-compat readers).
-    // Emit a real Rust `match` when present; else fall through to the flat
-    // list.
-    if let Some((scrutinee, arms, default)) = body.stmts.iter().find_map(|st| match st {
-        crate::mir::Stmt::Branch {
-            scrutinee,
-            arms,
-            default,
-        } => Some((scrutinee, arms, default)),
-        _ => None,
-    }) {
-        let scrutinee_rust = match scrutinee {
-            crate::mir::BranchScrutinee::Match(e) => e.rust.as_str(),
-            crate::mir::BranchScrutinee::Predicate(p) => p.0.rust.as_str(),
-        };
-        out.push_str(&format!("    match {} {{\n", scrutinee_rust));
-        let emit_arm_block = |out: &mut String, block: &crate::mir::Block| {
-            for (field, op_kind, value) in block_effect_triples(block) {
-                let field = field.as_str();
-                if account_env_struct.is_none() && field_type_is_pubkey(field, op, spec) {
-                    continue;
-                }
-                if account_env_struct.is_some() {
-                    emit_one_effect_with_account_env(
-                        out,
-                        op,
-                        spec,
-                        wrapping,
-                        field,
-                        op_kind,
-                        value,
-                        "            ",
-                        "accounts",
-                    );
-                } else {
-                    emit_one_effect(
-                        out,
-                        op,
-                        spec,
-                        wrapping,
-                        field,
-                        op_kind,
-                        value,
-                        "            ",
-                    );
-                }
-                emit_after_store_hooks(out, &mir.hooks, field, "            ");
-            }
-        };
-        for arm in arms {
-            let pattern = arm.pattern.as_ref().map(|p| p.rust.as_str()).unwrap_or("_");
-            out.push_str(&format!("        {} => {{\n", pattern));
-            emit_arm_block(out, &arm.block);
-            out.push_str("        }\n");
-        }
-        if let Some(default_block) = default {
-            out.push_str("        _ => {\n");
-            emit_arm_block(out, default_block);
-            out.push_str("        }\n");
-        } else {
-            // Spec patterns are literal-only, so synthesize a no-op
-            // wildcard to keep the match exhaustive even if the spec
-            // forgot the catch-all. The drift hash still records the
-            // spec's actual arms.
-            out.push_str("        _ => {}\n");
-        }
-        out.push_str("    }\n");
-    } else {
-        // #66 — iterate the lowered MIR body, not `op.effects`:
-        // `stmt_effect_triple` projects effect-shaped stmts onto the
-        // triple these templates consume (byte-identical; see its doc)
-        // and skips non-effect variants in-stream without reordering.
-        for (field, op_kind, value) in block_effect_triples(body) {
-            let field = field.as_str();
-            if account_env_struct.is_none() && field_type_is_pubkey(field, op, spec) {
-                continue;
-            }
-            if account_env_struct.is_some() {
-                emit_one_effect_with_account_env(
-                    out, op, spec, wrapping, field, op_kind, value, "    ", "accounts",
-                );
-            } else {
-                emit_one_effect(out, op, spec, wrapping, field, op_kind, value, "    ");
-            }
-            emit_after_store_hooks(out, &mir.hooks, field, "    ");
-        }
-    }
-
-    // Post-status assignment — drives the lifecycle transition declared in
-    // the handler signature (`State.X -> State.Y`). Combined with the pre-
-    // status check above, this turns lifecycle-only handlers into real
-    // state machines instead of `fn h() -> bool { true }` stubs.
-    if has_lifecycle(spec) {
-        if let Some(ref post) = op.post_status {
-            out.push_str(&format!("    s.status = Status::{};\n", post));
-        }
-    }
-
-    // Ghost (spec-only) field updates: a ghost with `on <this handler>`
-    // assigns after the normal effects; others are framed (unchanged).
-    // Values read `s.<ghost>` + params, matching the Lean transition.
-    // Arithmetic wraps in release (the `verify --proptest` path), so an
-    // arbitrary-state aggregate never panics on model overflow.
-    for ghost in &spec.ghosts {
-        for u in &ghost.updates {
-            if u.handler == op.name {
-                out.push_str(&format!("    s.{} = {};\n", ghost.name, u.value_rust));
-            }
-        }
-    }
-
-    out.push_str("    true\n");
-    out.push_str("}\n\n");
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::chumsky_adapter::parse_str;
-
-    /// #66 adaptor invariant — for every handler, the effect triples
-    /// projected from the lowered MIR body must equal
-    /// `ParsedHandler.effects` exactly (order + content). This is what
-    /// makes the `Stmt`-driven Kani/proptest emission byte-identical to
-    /// the former `op.effects` iteration. Checked over a synthetic spec
-    /// covering all seven op kinds plus the non-effect statement shapes.
-    #[test]
-    fn stmt_effect_triples_round_trip_parsed_effects() {
-        let src = r#"spec RoundTrip
-state {
-  pool : U64,
-  fees : U64,
-  spent : U64,
-  cap : U64,
-  owner : Pubkey,
-}
-type Error | Overflow | Unauthorized
-handler churn (amount : U64) (who : Pubkey) {
-  requires amount > 0 else Unauthorized
-  effect {
-    pool += amount,
-    fees +=! amount,
-    spent +=? amount,
-    cap -= amount,
-    pool -=! amount,
-    fees -=? amount,
-    owner := who,
-  }
-}
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mir = crate::mir::lower(&spec);
-        for op in &spec.handlers {
-            let body = mir.handler_block(&op.name).expect("mir body");
-            let got: Vec<(String, String, String)> = block_effect_triples(body)
-                .into_iter()
-                .map(|(f, k, v)| (f, k.to_string(), v.to_string()))
-                .collect();
-            let want: Vec<(String, String, String)> = op
-                .effects
-                .iter()
-                .map(|(f, k, v)| (f.clone(), k.clone(), v.clone()))
-                .collect();
-            assert_eq!(
-                got, want,
-                "MIR effect triples diverge from op.effects for `{}`",
-                op.name
-            );
-        }
-        // Sanity: the seven kinds all round-tripped (not vacuous).
-        let kinds: Vec<&str> = block_effect_triples(mir.handler_block("churn").expect("mir body"))
-            .iter()
-            .map(|(_, k, _)| *k)
-            .collect();
-        assert_eq!(
-            kinds,
-            vec!["add", "add_sat", "add_wrap", "sub", "sub_sat", "sub_wrap", "set"]
-        );
-    }
-
-    /// Phase-5 #42 — the transition emitter renders `effect { match … }`
-    /// from the lowered `Stmt::Branch` (per-arm semantics, wildcard arm
-    /// as `_`), pinned against the exact pre-Phase-5 `effect_branches`
-    /// output shape so the Kani/proptest harness text is unchanged.
-    #[test]
-    fn transition_fn_renders_branch_as_match() {
-        let src = r#"spec CondFee
-program_id "11111111111111111111111111111111"
-type State
-  | Active of { a : U64, b : U64, d : U64 }
-type Error
-  | InvalidAmount
-handler route (fee_type : U8) (amount : U64) : State.Active -> State.Active {
-  permissionless
-  requires amount > 0 else InvalidAmount
-  effect {
-    match fee_type {
-      0 => a +=! amount,
-      1 => b += amount,
-      _ => d := 0,
-    }
-  }
-}
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mir = crate::mir::lower(&spec);
-        let op = &spec.handlers[0];
-        let mut out = String::new();
-        emit_transition_fn(&mut out, &mir, op, &spec, false, |t| {
-            crate::codegen_shared::map_type(t, &spec)
-        })
-        .expect("emit");
-        let expected = "    match fee_type {\n\
-                        \x20       0 => {\n\
-                        \x20           s.a = s.a.saturating_add(amount);\n\
-                        \x20       }\n\
-                        \x20       1 => {\n\
-                        \x20           match s.b.checked_add(amount) {\n\
-                        \x20               Some(__v) => s.b = __v,\n\
-                        \x20               None => return false,\n\
-                        \x20           }\n\
-                        \x20       }\n\
-                        \x20       _ => {\n\
-                        \x20           s.d = 0;\n\
-                        \x20       }\n\
-                        \x20   }\n";
-        assert!(
-            out.contains(expected),
-            "transition must render the Branch as a Rust match:\n{out}"
-        );
-        // The flat union must NOT leak alongside the match.
-        assert_eq!(
-            out.matches("s.a = s.a.saturating_add(amount);").count(),
-            1,
-            "arm effect emitted exactly once:\n{out}"
-        );
-    }
-
-    /// Same invariant over every bundled example spec — the real-world
-    /// shapes (indexed records, variant prefixes, match-arm expansion,
-    /// transfers/CPI/emit interleaving).
-    #[test]
-    fn stmt_effect_triples_round_trip_bundled_examples() {
-        let examples = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .and_then(std::path::Path::parent)
-            .expect("repo root")
-            .join("examples/rust");
-        let mut checked = 0usize;
-        for entry in std::fs::read_dir(&examples).expect("examples/rust") {
-            let dir = entry.expect("dir entry").path();
-            if !dir.is_dir() {
-                continue;
-            }
-            let Some(spec_path) = std::fs::read_dir(&dir)
-                .expect("example dir")
-                .filter_map(|e| Some(e.ok()?.path()))
-                .find(|p| p.extension().is_some_and(|x| x == "qedspec"))
-            else {
-                continue;
-            };
-            let spec = match crate::check::parse_spec_file(&spec_path) {
-                Ok(s) => s,
-                // Brownfield onboarding fixtures may be intentionally
-                // partial — the invariant only applies to parseable specs.
-                Err(_) => continue,
-            };
-            let mir = crate::mir::lower(&spec);
-            for op in &spec.handlers {
-                let body = mir
-                    .handler_block(&op.name)
-                    .unwrap_or_else(|| panic!("MIR missing `{}` in {:?}", op.name, spec_path));
-                let got: Vec<(String, String, String)> = block_effect_triples(body)
-                    .into_iter()
-                    .map(|(f, k, v)| (f, k.to_string(), v.to_string()))
-                    .collect();
-                let want: Vec<(String, String, String)> = op
-                    .effects
-                    .iter()
-                    .map(|(f, k, v)| (f.clone(), k.clone(), v.clone()))
-                    .collect();
-                assert_eq!(
-                    got, want,
-                    "MIR effect triples diverge from op.effects for `{}` in {:?}",
-                    op.name, spec_path
-                );
-                checked += 1;
-            }
-        }
-        assert!(
-            checked > 10,
-            "expected to check many handlers, got {checked}"
-        );
-    }
-
-    #[test]
-    fn effect_target_base_strips_subscripts_and_dots() {
-        assert_eq!(effect_target_base("plain"), "plain");
-        assert_eq!(effect_target_base("accounts[i].active"), "accounts");
-        assert_eq!(effect_target_base("s.foo"), "s");
-        assert_eq!(effect_target_base("map[0]"), "map");
-        assert_eq!(effect_target_base("  padded  "), "padded");
-    }
-
-    #[test]
-    fn chained_subtraction_lowers_left_assoc_saturating() {
-        // `cap - used - reserved` is left-associative `Nat` subtraction in
-        // the Lean tier; the guard must lower recursive, correctly
-        // grouped, and saturating.
-        let got =
-            translate_guard_to_rust("amount <= state.cap - state.used - state.reserved", true);
-        assert_eq!(
-            got,
-            "amount <= s.cap.saturating_sub(s.used).saturating_sub(s.reserved)"
-        );
-        // Addition keeps left-assoc grouping (wrapping unchanged).
-        assert_eq!(
-            translate_guard_to_rust("state.a + state.b + state.c <= state.d", true),
-            "s.a.wrapping_add(s.b).wrapping_add(s.c) <= s.d"
-        );
-    }
-
-    #[test]
-    fn emit_transition_fn_default_add_emits_checked() {
-        // `pool += amount` defaults to checked semantics — overflow
-        // short-circuits via `return false`, matching deployed
-        // `checked_add(..).ok_or(err)?`.
-        let src = r#"spec T
-state { pool : U64 }
-handler buy (amount : U64) { effect { pool += amount } }
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mir = crate::mir::lower(&spec);
-        let op = &spec.handlers[0];
-        let mut out = String::new();
-        emit_transition_fn(&mut out, &mir, op, &spec, false, |t| {
-            crate::codegen_shared::map_type(t, &spec)
-        })
-        .expect("emit");
-        assert!(
-            out.contains("checked_add(amount)"),
-            "default `+=` should emit checked_add: {out}"
-        );
-        assert!(
-            out.contains("None => return false"),
-            "checked should short-circuit on None: {out}"
-        );
-        assert!(
-            !out.contains("wrapping_add") && !out.contains("saturating_add"),
-            "default `+=` should NOT emit wrapping/saturating: {out}"
-        );
-    }
-
-    #[test]
-    fn emit_transition_fn_saturating_add_emits_saturating() {
-        let src = r#"spec T
-state { pool : U64 }
-handler buy (amount : U64) { effect { pool +=! amount } }
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mir = crate::mir::lower(&spec);
-        let op = &spec.handlers[0];
-        let mut out = String::new();
-        emit_transition_fn(&mut out, &mir, op, &spec, false, |t| {
-            crate::codegen_shared::map_type(t, &spec)
-        })
-        .expect("emit");
-        assert!(
-            out.contains("saturating_add(amount)"),
-            "`+=!` should emit saturating_add: {out}"
-        );
-        assert!(
-            !out.contains("checked_add") && !out.contains("wrapping_add"),
-            "`+=!` should NOT emit checked/wrapping: {out}"
-        );
-    }
-
-    #[test]
-    fn emit_transition_fn_wrapping_add_emits_wrapping() {
-        let src = r#"spec T
-state { pool : U64 }
-handler buy (amount : U64) { effect { pool +=? amount } }
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mir = crate::mir::lower(&spec);
-        let op = &spec.handlers[0];
-        let mut out = String::new();
-        emit_transition_fn(&mut out, &mir, op, &spec, false, |t| {
-            crate::codegen_shared::map_type(t, &spec)
-        })
-        .expect("emit");
-        assert!(
-            out.contains("wrapping_add(amount)"),
-            "`+=?` should emit wrapping_add: {out}"
-        );
-        assert!(
-            !out.contains("checked_add") && !out.contains("saturating_add"),
-            "`+=?` should NOT emit checked/saturating: {out}"
-        );
-    }
-
-    #[test]
-    fn emit_transition_fn_sub_three_tiers() {
-        // Mirror: `-=` / `-=!` / `-=?` emit checked / saturating / wrapping.
-        for (op_str, expected) in &[
-            ("-=", "checked_sub(amount)"),
-            ("-=!", "saturating_sub(amount)"),
-            ("-=?", "wrapping_sub(amount)"),
-        ] {
-            let src = format!(
-                "spec T\nstate {{ pool : U64 }}\nhandler buy (amount : U64) {{ effect {{ pool {op_str} amount }} }}\n"
-            );
-            let spec = parse_str(&src).expect("parse");
-            let mir = crate::mir::lower(&spec);
-            let op = &spec.handlers[0];
-            let mut out = String::new();
-            emit_transition_fn(&mut out, &mir, op, &spec, false, |t| {
-                crate::codegen_shared::map_type(t, &spec)
-            })
-            .expect("emit");
-            assert!(
-                out.contains(expected),
-                "`{op_str}` should emit {expected}:\n{out}"
-            );
-        }
-    }
-
-    #[test]
-    fn emit_transition_fn_lifecycle_emits_status_guard_and_assignment() {
-        // Spec with a multi-state lifecycle. `transition` declares `Open ->
-        // Closed`, so the generated transition fn must (1) reject when the
-        // current status isn't `Open`, and (2) write `Status::Closed` on
-        // success. Without these, lifecycle-only handlers compile to
-        // `fn h() -> bool { true }` and every cover/liveness harness
-        // against them passes vacuously.
-        let src = r#"spec T
-type State
-  | Open of { x : U64 }
-  | Closed
-handler close : State.Open -> State.Closed { effect { x := 0 } }
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mir = crate::mir::lower(&spec);
-        let op = &spec.handlers[0];
-        let mut out = String::new();
-        emit_transition_fn(&mut out, &mir, op, &spec, false, |t| {
-            crate::codegen_shared::map_type(t, &spec)
-        })
-        .expect("emit");
-        assert!(
-            out.contains("if s.status != Status::Open"),
-            "lifecycle handler must reject when status mismatches pre_status:\n{out}"
-        );
-        assert!(
-            out.contains("s.status = Status::Closed;"),
-            "lifecycle handler must drive post_status assignment:\n{out}"
-        );
-    }
-
-    #[test]
-    fn emit_transition_fn_no_lifecycle_skips_status_lines() {
-        // Spec without a multi-state lifecycle (single State variant or
-        // flat record). emit_transition_fn must NOT emit any status guard
-        // or assignment — there's no Status enum to reference.
-        let src = r#"spec T
-state { balance : U64 }
-handler deposit (amount : U64) { effect { balance += amount } }
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mir = crate::mir::lower(&spec);
-        let op = &spec.handlers[0];
-        let mut out = String::new();
-        emit_transition_fn(&mut out, &mir, op, &spec, false, |t| {
-            crate::codegen_shared::map_type(t, &spec)
-        })
-        .expect("emit");
-        assert!(
-            !out.contains("Status::"),
-            "lifecycle-free spec must not reference Status:\n{out}"
-        );
-    }
-
-    #[test]
-    fn emit_state_struct_appends_status_when_lifecycle_present() {
-        let src = r#"spec T
-type State
-  | Open of { x : U64 }
-  | Closed
-handler close : State.Open -> State.Closed { effect { x := 0 } }
-"#;
-        let spec = parse_str(src).expect("parse");
-        let mutable = mutable_fields(&spec.state_fields);
-        let mut out = String::new();
-        emit_state_struct_with_lifecycle(
-            &mut out,
-            &mutable,
-            "Clone, Copy",
-            |t| Ok(t.to_string()),
-            has_lifecycle(&spec),
-        )
-        .expect("emit");
-        assert!(
-            out.contains("status: Status,"),
-            "lifecycle spec must inject `status: Status` field:\n{out}"
-        );
-    }
-
-    #[test]
-    fn kani_pubkey_rewrite_handles_account_and_state_fields() {
-        let src = r#"spec T
-type State | Active of { admin_key : Pubkey }
-type Error | Unauthorized
-handler set_admin : State.Active -> State.Active {
-  accounts { admin : signer }
-  requires admin.pubkey == state.admin_key else Unauthorized
-  effect { admin_key := admin.pubkey }
-}
-"#;
-        let spec = parse_str(src).expect("parse");
-        let op = &spec.handlers[0];
-        let expr = "(accounts.admin.pubkey == s.admin_key) && (amount > 0)";
-        let rewritten = rewrite_kani_pubkey_comparisons(expr, op, &spec);
-        assert_eq!(
-            rewritten,
-            "(pubkey_eq(&accounts.admin.pubkey, &s.admin_key)) && (amount > 0)"
-        );
-    }
-
-    #[test]
-    fn kani_pubkey_rewrite_handles_indexed_pubkey_arrays() {
-        let src = r#"spec T
-const MAX_MEMBERS = 32
-type State | Active of { members : Map[MAX_MEMBERS] Pubkey }
-handler approve (member_index : U8) (approver : Pubkey) : State.Active -> State.Active {
-  requires state.members[member_index] == approver
-}
-"#;
-        let spec = parse_str(src).expect("parse");
-        let op = &spec.handlers[0];
-        let expr = "(s.members[(member_index) as usize] == approver) && (member_index < 32)";
-        let rewritten = rewrite_kani_pubkey_comparisons(expr, op, &spec);
-        assert_eq!(
-            rewritten,
-            "(pubkey_eq(&s.members[(member_index) as usize], &approver)) && (member_index < 32)"
-        );
-    }
-
-    #[test]
-    fn split_top_level_and_splits_only_balanced_top_level_terms() {
-        assert_eq!(
-            split_top_level_and("amount > 0 && fee_bps <= 100 && min_out > 0"),
-            vec!["amount > 0", "fee_bps <= 100", "min_out > 0"]
-        );
-        assert_eq!(
-            split_top_level_and("(amount > 0 && fee_bps <= 100) && min_out > 0"),
-            vec!["(amount > 0 && fee_bps <= 100)", "min_out > 0"]
-        );
-        assert_eq!(
-            split_top_level_and(
-                "is_allowed(mints[(lane) as usize] == mint && lane < 32) && amount > 0"
-            ),
-            vec![
-                "is_allowed(mints[(lane) as usize] == mint && lane < 32)",
-                "amount > 0"
-            ]
-        );
-    }
-
-    #[test]
-    fn collect_guard_terms_splits_guard_and_requires_without_nested_or_splits() {
-        let src = r#"spec T
-type State | Active of { admin_key : Pubkey, allowed : Bool }
-type Error | Unauthorized | InvalidAmount
-handler swap (amount : U64) (min_out : U64) : State.Active -> State.Active {
-  accounts { admin : signer }
-  requires admin.pubkey == state.admin_key else Unauthorized
-  requires amount >= min_out and min_out > 0 else InvalidAmount
-}
-"#;
-        let mut spec = parse_str(src).expect("parse");
-        spec.handlers[0].guard_str = Some("state.allowed && (amount > 0 || min_out > 0)".into());
-        let op = &spec.handlers[0];
-        let terms = collect_guard_terms_with_account_env(op, false, Some("accounts"));
-        let exprs = terms
-            .into_iter()
-            .map(|term| term.rust_expr)
-            .collect::<Vec<_>>();
-        assert_eq!(
-            exprs,
-            vec![
-                "s.allowed",
-                "(amount > 0 || min_out > 0)",
-                "accounts.admin.pubkey == s.admin_key",
-                "(amount >= min_out)",
-                "(min_out > 0)",
-            ]
-        );
-    }
-
-    #[test]
-    fn rewrite_kani_bps_mul_div_uses_solver_friendly_helper() {
-        assert_eq!(
-            rewrite_kani_bps_mul_div(
-                "fee_output_normalized >= (fee_input_normalized * retained_value_bps) / 10000"
-            ),
-            "fee_output_normalized >= mul_bps_floor_u128(fee_input_normalized, retained_value_bps)"
-        );
-        assert_eq!(
-            rewrite_kani_bps_mul_div("amount_in * fee_bps / 10000 <= amount_in"),
-            "mul_bps_floor_u128(amount_in, fee_bps) <= amount_in"
-        );
-        assert_eq!(
-            rewrite_kani_bps_mul_div("(a + b) * fee_bps / 10000 <= a"),
-            "(a + b) * fee_bps / 10000 <= a"
-        );
-    }
-
-    #[test]
-    fn rewrite_kani_checked_add_equality_avoids_overflow_checks() {
-        assert_eq!(
-            rewrite_kani_guard_arithmetic("max_fee_bps + retained_value_bps == 10000"),
-            "max_fee_bps.checked_add(retained_value_bps) == Some(10000)"
-        );
-        assert_eq!(
-            rewrite_kani_guard_arithmetic(
-                "fee_output_normalized >= (fee_input_normalized * retained_value_bps) / 10000 && max_fee_bps + retained_value_bps == 10000"
-            ),
-            "fee_output_normalized >= mul_bps_floor_u128(fee_input_normalized, retained_value_bps) && max_fee_bps.checked_add(retained_value_bps) == Some(10000)"
-        );
-    }
-
-    #[test]
-    fn negate_simple_top_level_comparison_flips_only_outer_operator() {
-        assert_eq!(
-            negate_simple_top_level_comparison(
-                "fee_output_normalized >= mul_bps_floor_u128(fee_input_normalized, retained_value_bps)"
-            ),
-            Some(
-                "fee_output_normalized < mul_bps_floor_u128(fee_input_normalized, retained_value_bps)"
-                    .to_string()
-            )
-        );
-        assert_eq!(
-            negate_simple_top_level_comparison(
-                "(mul_bps_floor_u128(amount_in, fee_bps) <= amount_in)"
-            ),
-            Some("mul_bps_floor_u128(amount_in, fee_bps) > amount_in".to_string())
-        );
-        assert_eq!(
-            negate_simple_top_level_comparison(
-                "pubkey_eq(&accounts.input_mint.pubkey, &s.allowed_mint_0) || amount > 0"
-            ),
-            None
-        );
-    }
-
-    #[test]
-    fn check_effect_targets_accepts_declared_fields() {
-        let src = r#"spec T
-state { balance : U64 }
-handler deposit (amount : U64) {
-  effect { balance += amount }
-}"#;
-        let spec = parse_str(src).expect("parse");
-        assert!(check_effect_targets(&spec).is_ok());
-    }
-
-    #[test]
-    fn check_effect_targets_errors_on_undeclared_target() {
-        // Effect writes `phantom` but the state declares only `balance`;
-        // the error must name the handler and the bad field.
-        let src = r#"spec T
-state { balance : U64 }
-handler bogus (amount : U64) {
-  effect { phantom := amount }
-}"#;
-        let spec = parse_str(src).expect("parse");
-        let err = check_effect_targets(&spec).unwrap_err().to_string();
-        assert!(err.contains("bogus"), "should name handler: {err}");
-        assert!(err.contains("phantom"), "should name field: {err}");
-    }
-}
